Backend/MusicStream/requirements.txt
Django==3.1.12
djangorestframework==3.13.1
djangorestframework_jwt==1.11.0
